// database/contract_milestones_data/types.ts
export interface ContractMilestoneData {
  contract_id: number;
  milestone_sequential_number: number;
  milestone_criteria: string | null;
  milestone_due_date: string | null;
  milestone_amount: number;
  milestone_percentage: number;
  progress_status: number;
  milestone_condition_met_indicator: number;
  progress_status_date: string | null;
  milestone_met_date: string | null;
  progress_status_user_id: number | null;
  milestone_met_mark_user_id: number | null;
}

export interface CreateContractMilestoneInput {
  contract_id: number;
  milestone_sequential_number: number | null;
  milestone_criteria: string | null;
  milestone_due_date: string | null;
  milestone_amount: number | null;
  milestone_percentage: number | null;
  progress_status: number | null;
  milestone_condition_met_indicator: number | null;
  progress_status_date: string | null;
  milestone_met_date: string | null;
  progress_status_user_id: number | null;
  milestone_met_mark_user_id: number | null;
}

export interface UpdateContractMilestoneInput {
  milestone_criteria?: string | null;
  milestone_due_date?: string | null;
  milestone_amount?: number | null;
  milestone_percentage?: number | null;
  progress_status?: number | null;
  milestone_condition_met_indicator?: number | null;
  progress_status_date?: string | null;
  milestone_met_date?: string | null;
  progress_status_user_id?: number | null;
  milestone_met_mark_user_id?: number | null;
}